feat(governance): configure Robinhood deployment controls
- Reconstruct the verified source state from 1bb7532 through 86ff390
- Preserve production behavior and behavior-focused regression coverage
- Publish configurable delays and testnet deployment preparation

+++ src/types/MLOProfitShareTypes.sol
@@ -11,8 +11,8 @@ library MLOProfitShareTypes {
 
     struct PendingProfitSplit {
         ProfitSplit split;
-        uint64 executableAt;
-        uint64 expiresAt;
+        uint256 executableAt;
+        uint256 expiresAt;
         bool exists;
     }
